refactor nav bar logic
