+++ app/src/Navigation/account.ts
@@ -1,6 +1,6 @@
 import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 
-import HomeNavBar from '../components/HomeNavBar';
+import HomeNavBar from '../components/NavBar/HomeNavBar';
 import DisclaimerScreen from '../screens/DisclaimerScreen';
 import HomeScreen from '../screens/HomeScreen';
 import { black } from '../utils/colors';
