+++ app/src/components/NavBar/index.ts
@@ -1,0 +1,2 @@
+export * from './DefaultNavBar';
+export * from './HomeNavBar';
